requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-CORS
pymysql
cryptography
email-validator